move hand_coded_optimizations to heuristic.py [pr]
also folded all long lines
