+++ test/external/speed_compare_cuda_nv.py
@@ -2,6 +2,7 @@ from tinygrad import Device, dtypes
 from tinygrad.helpers import getenv, colorize_float
 from extra.optimization.helpers import load_worlds, ast_str_to_lin
 from test.external.fuzz_linearizer import get_fuzz_rawbufs
+from tinygrad.codegen.heuristic import hand_coded_optimizations
 from tinygrad.engine.search import bufs_from_lin
 from tinygrad.engine.realize import CompiledRunner
 from tinygrad.tensor import _to_np_dtype
@@ -21,7 +22,7 @@ if __name__ == "__main__":
   for num,ast in enumerate(ast_strs):
     # cuda compile
     culin = ast_str_to_lin(ast, opts=cudev.renderer)
-    culin.hand_coded_optimizations()
+    culin = hand_coded_optimizations(culin)
     has_bf16 = any(b.dtype == dtypes.bfloat16 for b in culin.membufs)
 
     cuda_prg = CompiledRunner(culin.to_program())
@@ -31,7 +32,7 @@ if __name__ == "__main__":
     rdr = nvdev.renderer
     rdr.device = "NV"
     nvlin = ast_str_to_lin(ast, opts=rdr)
-    nvlin.hand_coded_optimizations()
+    nvlin = hand_coded_optimizations(nvlin)
     nv_prg = CompiledRunner(nvlin.to_program())
     nvbufs = bufs_from_lin(nvlin)
     test_nvbufs = get_fuzz_rawbufs(nvlin) if not has_bf16 else nvbufs
