+++ test/external/speed_compare_cuda_ptx.py
@@ -1,6 +1,7 @@
 import itertools
 from tinygrad import Device
 from tinygrad.engine.realize import CompiledRunner
+from tinygrad.codegen.heuristic import hand_coded_optimizations
 from tinygrad.helpers import getenv, colorize_float
 from extra.optimization.helpers import load_worlds, ast_str_to_lin
 from tinygrad.engine.search import bufs_from_lin
@@ -23,7 +24,7 @@ if __name__ == "__main__":
     # cuda compile
     dev.compiler = CUDACompiler(dev.arch)
     lin = ast_str_to_lin(ast, opts=dev.renderer)
-    lin.hand_coded_optimizations()
+    lin = hand_coded_optimizations(lin)
     cuda_prg = CompiledRunner(lin.to_program())
 
     bufs = bufs_from_lin(lin)
@@ -31,7 +32,7 @@ if __name__ == "__main__":
     # ptx compile
     dev.compiler = PTXCompiler(dev.arch)
     lin = ast_str_to_lin(ast, opts=ptx)
-    lin.hand_coded_optimizations()
+    lin = hand_coded_optimizations(lin)
     lin.linearize()
     ptx_prg = CompiledRunner(lin.to_program())
